requirements.txt
Flask
flask-cors
pymongo
python-dotenv
openai
instagrapi
spotipy
schedule
twilio
selenium